Update medicine
Treat each other Well.
This is one of the last chances we have to show solidarity and kindness in agape. This is the first time we have made it to 2025.

Thats cause I stole billions of hours' time back from  Cassiopia and heaven, You cannot live without time, We Should have the same rules that we do for water, food, shelter/ When Your cared for: It makes your heart sing

Everytime someone does something for you out of their love.
Both The Gifted and the Gifter score magic love points.

than walk hand in hand into the sunset of symbosis. forever.